% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{blackBodyRadiationColors}
\alias{blackBodyRadiationColors}
\title{blackBodyRadiationColors}
\usage{
blackBodyRadiationColors(x, max_value = 1)
}
\arguments{
\item{x}{numeric should be between 0 (black) and 1 (white)}

\item{max_value}{numeric maximum value to be used for scaling}
}
\value{
returns no value, but has side effect of generating color map
}
\description{
set up color scale for black body spectrum
}
\details{
I obtained this by copy and paste from internet (reference unknown)
}
\examples{
colors.blackBody <- rev(blackBodyRadiationColors(seq(0.3,1,length.out=20)))

}
