% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{catNum3}
\alias{catNum3}
\title{catNum3}
\usage{
catNum3(l)
}
\arguments{
\item{l}{list each component corresponds to an ontology branch, and contains the decreasing sorted
tabulation of output of the number of times that a category appears in a triplet}
}
\value{
returns a list each component corresponds to an ontology branch, and contains a vector of category plotting positions
}
\description{
assign the axis coordinate number to be used for the plotting position for each category
}
\details{
a component of l is like:
GO_0005515__protein_binding GO_0042802__identical_protein_binding   GO_0005178__integrin_binding
38                          4                                        3

#' a component of l1 is like:
GO_0005515__protein_binding GO_0042802__identical_protein_binding  GO_0005178__integrin_binding
1                           2                                      3
}
\examples{
#load("data/x_l.RData")
catNum3(x_l)

}
