% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{graphIt}
\alias{graphIt}
\title{graphIt}
\usage{
graphIt(mat3d, sp, w, r, verbose)
}
\arguments{
\item{mat3d}{component of HTGM3Ddriver() output list}

\item{sp}{integer vector containing c(x,y,z) coordinated of point}

\item{w}{integer line number within mat3d}

\item{r}{numeric max value of x,y,z ranges}

\item{verbose}{integer vector representing vprint classes}
}
\value{
returns no value, but has side effect of annotating the 3D graph
}
\description{
annotate a selected point in the 3d graphic
}
