% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{insertCatSize}
\alias{insertCatSize}
\title{insertCatSize}
\usage{
insertCatSize(mat, GOGOA3)
}
\arguments{
\item{mat}{return value of Jaccard3()}

\item{GOGOA3}{return value of subsetGOGOA()}
}
\value{
returns augmented version of matrix mat
}
\description{
compute fraction of total genes in the entire ontology that map to each category,
and insert column into matrix mat
}
\examples{
\dontrun{
# GOGOA3.RData is too large to include in the R package
# so I need to load it from a file that is not in the package.
# Since this is in a file in my own file system, I could not
# include this as a regular example in the package.
# you can generate it using the package 'minimalistGODB'
# or you can retrieve it from https://github.com/barryzee/GO/tree/main/databases
load("/Users/barryzeeberg/personal/GODB_RDATA/goa_human/GOGOA3_goa_human.RData")

mat2<-insertCatSize(x_mat,GOGOA3)
}

}
