% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{interactWithGraph3D}
\alias{interactWithGraph3D}
\title{interactWithGraph3D}
\usage{
interactWithGraph3D(mat3d, maxfract = 1, newWindow = TRUE, verbose = TRUE)
}
\arguments{
\item{mat3d}{component of HTGM3Ddriver() output list}

\item{maxfract}{numeric upper threshold for category size to display}

\item{newWindow}{Boolean if TRUE open new window to avoid over writing current window}

\item{verbose}{integer vector representing vprint classes}
}
\value{
returns no value, but has side effect of annotating the 3D graph
}
\description{
rotate the 3d graph and/or select a point within the 3D graph, and annotate that point
}
\examples{
if(interactive()){
#load("data/x_mat3d.RData")
interactWithGraph3D(x_mat3d)
}

}
