% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{Jaccard3}
\alias{Jaccard3}
\title{Jaccard3}
\usage{
Jaccard3(dir, m, thresh1 = 3, thresh3 = 3, verbose = 2)
}
\arguments{
\item{dir}{character string full path name to the directory acting as result repository}

\item{m}{return value of pruneCatGenes()}

\item{thresh1}{parameter passed to Jaccard3()}

\item{thresh3}{parameter passed to Jaccard3()}

\item{verbose}{integer vector representing vprint classes}
}
\value{
returns matrix tabulating genes in the intersection of categories from 3 ontology branches
also has side effect of saving files containing those genes
}
\description{
compute the number of genes in the intersection of categories from 3 ontology branches
}
\examples{
#load("data/x_m11.RData")
Jaccard3(tempdir(),x_m11,thresh1=3,thresh3=3,verbose=1:5)

}
