% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{plot3Dmat}
\alias{plot3Dmat}
\title{plot3Dmat}
\usage{
plot3Dmat(mat, l)
}
\arguments{
\item{mat}{return value of HTGM3D()}

\item{l}{return value of catNum3()}
}
\value{
augmented version of matrix containing x,y,z coordinates for each triplet
}
\description{
compute x,y,z coordinates for each triplet
}
\examples{
#load("data/x_mat.RData")
#load("data/x_l.RData")
p3<-plot3Dmat(x_mat,x_l)

}
