% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{pruneCatGenes}
\alias{pruneCatGenes}
\title{pruneCatGenes}
\usage{
pruneCatGenes(m, mn = 2, mx = 200)
}
\arguments{
\item{m}{the return value of catGenes()}

\item{mn}{integer min category size threshold passed to trimGOGOA3()}

\item{mx}{integer max category size threshold passed to trimGOGOA3()}
}
\value{
returns pruned version of matrix m
}
\description{
eliminate those categories to which no genes map
}
\examples{
#load("data/x_cg.RData")
m<-pruneCatGenes(x_cg,2,200)

}
