% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{showGenes}
\alias{showGenes}
\title{showGenes}
\usage{
showGenes(mat3d, w, range, npad)
}
\arguments{
\item{mat3d}{component of HTGM3Ddriver() output list}

\item{w}{integer line number within mat3d}

\item{range}{list of ranges}

\item{npad}{integer number of blanks for padding}
}
\value{
returns no value, but has side effect of opening textEdit
}
\description{
open gene list in a textEdit window
}
