% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{subMatDiffs}
\alias{subMatDiffs}
\title{subMatDiffs}
\usage{
subMatDiffs(m1, m2, verbose = 3)
}
\arguments{
\item{m1}{matrix}

\item{m2}{matrix}

\item{verbose}{integer vector representing vprint classes}
}
\value{
returns no values
}
\description{
retrieve submatrices of common rownames and colnames, and report differences
}
\details{
compare submatrices of m1 and m2 that have common rownames and colnames
}
\examples{
rn<-c("a","b","c")
cn<-c("a","b","c","d")
m1<-matrix(1:12,nrow=length(rn),ncol=length(cn))
rownames(m1)<-rn
colnames(m1)<-cn
m2<-m1
subMatDiffs(m1,m2)

m3<-m1
m3[1,1]<-0
m3[1,2]<-0
subMatDiffs(m1,m3)

m4<-m3
colnames(m4)<-c("aa","b","c","d")
subMatDiffs(m1,m4)

}
