% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM3D.R
\name{whichOS}
\alias{whichOS}
\title{whichOS}
\usage{
whichOS()
}
\value{
"BAD" Tahoe 26.2, "OK" otherwise
}
\description{
determine if host machine is running OSX Tahoe 26.2
}
\details{
as of Feb 6 2026 Tahoe 26.2 does not interact properly with rgl Xwindows package
}
